refactor(canon): keep the corsa CLI files under the source length policy
